feat: make the Borrow market screen responsive
