+++ frontend/src/app/app/LendMarket/lendersMarket.tsx
@@ -254,11 +254,10 @@ const Pagination = ({
       {Array.from({ length: totalPages }, (_, index) => (
         <button
           key={index}
-          className={`px-4 py-2 ${
-            currentPage === index + 1
+          className={`px-4 py-2 ${currentPage === index + 1
               ? "bg-[rgba(0,0,0,0.8)] text-white"
               : "bg-[#F5F5F5] text-black border-black border"
-          } rounded-lg`}
+            } rounded-lg`}
           onClick={() => onPageChange(index + 1)}
         >
           {index + 1}
@@ -363,4 +362,3 @@ const Lender = () => {
 };
 
 export default Lender;
-
