+++ frontend/src/app/app/LendMarket/page.tsx
@@ -1,11 +1,11 @@
 import Lender from "./lendersMarket";
 
 const Page = () => {
-    return (
-        <div className="bg-[#F5F5F5]">
-        <Lender/>
-        </div>
-     );
+  return (
+    <div className="bg-[#F5F5F5]">
+      <Lender />
+    </div>
+  );
 }
- 
+
 export default Page;
